trading_bot/requirements.txt
aiohttp
python-telegram-bot
python-dotenv
pandas
pytest
pytest-asyncio
cryptography==40.0.2
websockets
